+++ site/dat/docs/Changelog.md
@@ -2,9 +2,12 @@
 
 ## 1.6.3 (next)
  - fix percentile value handling in passfail criteria
+ - add setting up of CWD on server side
  - fix problem with sending included configs into cloud
  - fix cumulative fail criteria processing order
  - add extraction of `scenario` dictionary into `scenarios` section, remove inlined scenarios mentions from docs
+ - limit the amount of monitoring data BlazeMeter uploader accumulates
+ - fix passfail criteria with no last datapoint available
 
 ## 1.6.2 <sup>8 jun 2016</sup>
  - fix passfail-related regression crash